ekf2: move rng height to file
