+++ src/modules/ekf2/EKF/control.cpp
@@ -654,102 +654,6 @@ void Ekf::controlGpsYawFusion(bool gps_checks_passing, bool gps_checks_failing)
 	}
 }
 
-bool Ekf::isConditionalRangeAidSuitable()
-{
-	bool is_range_aid_suitable = false;
-
-	if (_control_status.flags.in_air
-	    && _range_sensor.isHealthy()
-	    && isTerrainEstimateValid()) {
-		// check if we can use range finder measurements to estimate height, use hysteresis to avoid rapid switching
-		// Note that the 0.7 coefficients and the innovation check are arbitrary values but work well in practice
-		const float range_hagl = _terrain_vpos - _state.pos(2);
-		const float range_hagl_max = _control_status.flags.rng_hgt ? _params.max_hagl_for_range_aid : (_params.max_hagl_for_range_aid * 0.7f);
-		const bool is_in_range = range_hagl < range_hagl_max;
-
-		const float hagl_test_ratio = (_hagl_innov * _hagl_innov / (sq(_params.range_aid_innov_gate) * _hagl_innov_var));
-		const bool is_hagl_stable = _control_status.flags.rng_hgt ? (hagl_test_ratio < 1.f) : (hagl_test_ratio < 0.01f);
-
-		if (isHorizontalAidingActive()) {
-			const float max_vel = _control_status.flags.rng_hgt ? _params.max_vel_for_range_aid : (_params.max_vel_for_range_aid * 0.7f);
-			const bool is_below_max_speed = !_state.vel.xy().longerThan(max_vel);
-
-			is_range_aid_suitable = is_in_range && is_hagl_stable && is_below_max_speed;
-
-		} else {
-			is_range_aid_suitable = is_in_range && is_hagl_stable;
-		}
-
-	}
-
-	return is_range_aid_suitable;
-}
-
-void Ekf::controlRangeHeightFusion()
-{
-	if (!((_params.rng_ctrl == RngCtrl::CONDITIONAL) || (_params.rng_ctrl == RngCtrl::ENABLED))) {
-		stopRngHgtFusion();
-		return;
-	}
-
-	_rng_hgt_b_est.predict(_dt_ekf_avg);
-
-	const bool rng_intermittent = !isRecent(_time_last_range, 2 * RNG_MAX_INTERVAL);
-
-	// If we are supposed to be using range finder data as the primary height sensor, have bad range measurements
-	// and are on the ground, then synthesise a measurement at the expected on ground value
-	if (!_control_status.flags.in_air
-	    && !_range_sensor.isDataHealthy()
-	    && _range_sensor.isRegularlySendingData()
-	    && _range_sensor.isDataReady()) {
-
-		_range_sensor.setRange(_params.rng_gnd_clearance);
-		_range_sensor.setValidity(true); // bypass the checks
-	}
-
-	if (_rng_data_ready) {
-		updateRngHgt(_aid_src_rng_hgt);
-
-		const bool do_conditional_range_aid = (_params.rng_ctrl == RngCtrl::CONDITIONAL) && isConditionalRangeAidSuitable();
-
-		const bool continuing_conditions_passing = _range_sensor.isDataHealthy() && !rng_intermittent
-							   && ((_params.rng_ctrl == RngCtrl::ENABLED) || do_conditional_range_aid);
-		const bool starting_conditions_passing = continuing_conditions_passing
-							 && _range_sensor.isRegularlySendingData();
-
-		if (_control_status.flags.rng_hgt) {
-			if (continuing_conditions_passing) {
-				fuseRngHgt(_aid_src_rng_hgt);
-
-				const bool is_fusion_failing = isTimedOut(_aid_src_rng_hgt.time_last_fuse, _params.hgt_fusion_timeout_max);
-
-				if (isHeightResetRequired()) {
-					// All height sources are failing
-					resetHeightToRng();
-					resetVerticalVelocityToZero();
-
-				} else if (is_fusion_failing) {
-					// Some other height source is still working
-					stopRngHgtFusion();
-					_control_status.flags.rng_fault = true;
-					_range_sensor.setFaulty();
-				}
-
-			} else {
-				stopRngHgtFusion();
-			}
-
-		} else {
-			if (starting_conditions_passing) {
-				startRngHgtFusion();
-			}
-		}
-
-	} else if (_control_status.flags.rng_hgt && rng_intermittent) {
-		stopRngHgtFusion();
-	}
-}
-
 void Ekf::controlAirDataFusion()
 {
 	// control activation and initialisation/reset of wind states required for airspeed fusion
